rename OnyxNavigationBar to OnyxNavBar
