+++ packages/sit-onyx/src/components/OnyxNavBar/types.ts
@@ -1,6 +1,6 @@
 import type { OnyxNavAppAreaProps } from "../OnyxNavAppArea/types";
 
-export type OnyxNavigationBarProps = OnyxNavAppAreaProps & {
+export type OnyxNavBarProps = OnyxNavAppAreaProps & {
   /**
    * Whether to show a back button.
    */
